Util (Util v1.0.0)

Módulo con funciones que se reutilizan

autor: Simon Lopez E, Luna Sofia Oviedo Rios, David Alejandro Henao Jaramillo

fecha: septiembre del 2026

licencia: GNU GPL v3

Functions

formatear_moneda(valor)

ingresar(mensaje)

Solicita un dato por consola y retorna el string limpio.

mostrar_mensaje(mensaje)

Función que muestra un mensaje en pantalla

Parametro

mensaje: string que se mostrará en pantalla

Ejemplo

iex> Util.mostrar_mensaje("Hola mundo")

o puede usar

"Hola mundo"
|> Util.mostrar_mensaje()