alias (util v1.1.5)

Creation of module aliases

Functions

create(Alias, OrigMod)

Create a module Alias for the given OrigMod module. This function is useful for creating in-memory Elixir wrapper modules without needing to maintain persistent files with such wrappers. IncFuns argument can be all or a list of {Fun, Arity} values which only export a subset of functions from the OrigMod.

create(Alias, OrigMod, IncFuns)