gen_leader behaviour (util v1.1.5)

Leader election behavior.
This application implements a leader election behavior modeled after gen_server. This behavior intends to make it reasonably straightforward to implement a fully distributed server with master-slave semantics.

The gen_leader behavior supports nearly everything that gen_server does (some functions, such as multicall() and the internal timeout, have been removed), and adds a few callbacks and API functions to support leader election etc.

Also included is an example program, a global dictionary, based on the modules gen_leader and dict. The callback implementing the global dictionary is called 'test_cb', for no particularly logical reason.

New version: The internal leader election algorithm was faulty and has been replaced with a new version based on a different leader election algorithm. As a consequence of this the query functions alive and down can no longer be provided. The new algorithm also make use of an incarnation parameter, by default written to disk in the function incarnation. This implies that only one gen_leader per node is permitted, if used in a diskless environment, incarnation must be adapted.

Modifications contributed by Serge Aleynikov:

Added configurable startup options (see leader_options() type)

Implemented handle_DOWN/3 callback with propagation of the leader's state via broadcast to all connected candidates.

Fixed population of the #election.down member so that down/1 query can be used in the behavior's implementation

Rewrote implementation of the tau timer to prevent the leader looping on the timer timeout event when all candidates are connected.

Types

bcast_type/0

-type bcast_type() :: all | sender.

caller_ref/0

-type caller_ref() :: {pid(), reference()}.

elid/0

-type elid() :: {priority(), incarn(), lclock()}.

incarn/0

-type incarn() :: non_neg_integer().

lclock/0

-type lclock() :: non_neg_integer().

mon_ref/0

-type mon_ref() :: reference().

name/0

-type name() :: atom().

option/0

-type option() ::
    {workers, Workers :: [node()]} |
    {vardir, Dir :: string()} |
    {bcast_type, Type :: bcast_type()} |
    {heartbeat, Seconds :: integer()}.

options/0

-type options() :: [option()].

priority/0

-type priority() :: integer().

server_ref/0

-type server_ref() :: name() | {name(), node()} | {global, name()} | pid().

start_ret/0

-type start_ret() :: {ok, pid()} | {error, term()}.

status/0

-type status() :: elec1 | elec2 | wait | joining | worker | waiting_worker | norm.

Functions

alive(Election)

-spec alive(#election{}) -> [node()].

Returns list of alive nodes.

behaviour_info(Other)

-spec behaviour_info(atom()) -> undefined | [{atom(), arity()}].

broadcast(_, ToNodes, E)

call(Name, Request)

-spec call(server_ref(), term()) -> term().

Equivalent to gen_server:call/2, but with a slightly different exit reason if something goes wrong. This function calls the gen_leader process exactly as if it were a gen_server (which, for practical purposes, it is.)

call(Name, Request, Timeout)

-spec call(server_ref(), term(), integer()) -> term().

Equivalent to gen_server:call/3, but with a slightly different exit reason if something goes wrong. This function calls the gen_leader process exactly as if it were a gen_server (which, for practical purposes, it is.)

candidates(Election)

-spec candidates(#election{}) -> [node()].

Returns a list of known candidates.

cast(Name, Request)

-spec cast(Name :: name() | pid(), Request :: term()) -> ok.

Equivalent to gen_server:cast / 2.

down(Election)

-spec down(#election{}) -> [node()].

Returns list of down nodes.

format_status(Opt, StatusData)

init_it(Starter, Parent, Name, Mod, _, Options)

leader_call(Name, Request)

-spec leader_call(Name :: server_ref(), Request :: term()) -> term().

Makes a call (similar to gen_server:call/2) to the leader. The call is forwarded via the local gen_leader instance, if that one isn't actually the leader. The client will exit if the leader dies while the request is outstanding.
This function uses gen:call/3, and is subject to the same default timeout as e.g. gen_server:call/2.

leader_call(Name, Request, Timeout)

-spec leader_call(Name :: server_ref(), Request :: term(), Timeout :: integer()) -> term().

Makes a call (similar to gen_server:call/3) to the leader. The call is forwarded via the local gen_leader instance, if that one isn't actually the leader. The client will exit if the leader dies while the request is outstanding.

leader_cast(Name, Request)

-spec leader_cast(Name :: name() | pid(), Request :: term()) -> ok.

Similar to gen_server:cast/2 but will be forwarded to the leader via the local gen_leader instance.

leader_node(Election)

-spec leader_node(#election{}) -> node() | none.

Returns the current leader node.

print_event(Dev, Event, Name)

reply(From, Reply)

-spec reply(From :: caller_ref(), Reply :: term()) -> term().

Equivalent to gen_server:reply / 2.

start(Name, CandidateNodes, OptArgs, Mod, Arg, Options)

-spec start(Name :: atom(),
      CandidateNodes :: [node()],
      OptArgs :: options(),
      Mod :: module(),
      Arg :: term(),
      Options :: list()) ->
         start_ret().

Starts a gen_leader process without linking to the parent.
See also: start_link/6.

start_link(Name, CandidateNodes, OptArgs, Mod, Arg, Options)

-spec start_link(Name :: atom(),
           CandidateNodes :: [node()],
           OptArgs :: options(),
           Mod :: module(),
           Arg :: term(),
           Options :: list()) ->
              start_ret().

Starts a gen_leader process. NameThe locally registered name of the processCandidateNodesThe names of nodes capable of assuming a leadership roleOptArgsOptional arguments given to gen_leader.
{workers, Workers}

The names of nodes that will be part of the "cluster", but cannot ever assume a leadership role. Default: [].

{vardir, Dir}

Directory name used to store candidate's incarnation cookie. Default: "."

{bcast_type, Type}

When Type is 'all' each election event (when a new candidate becomes visible to the leader) will be broadcast to all live candidate nodes. Each candidate will get a from_leader/3 callback. When Type is sender, only the newly registered candidate will get the surrendered/3 callback. Default: sender.

{heartbeat, Seconds}

Heartbeat timeout value used to send ping messages to inactive candidate nodes.
ModThe name of the callback moduleArgArgument passed on to Mod:init/1OptionsSame as gen_server's Options
The list of candidates needs to be known from the start. Workers could potentially be added at runtime, but no functionality to do this is provided by this version.

system_code_change(_, Module, OldVsn, Extra)

system_continue(Parent, Debug, _)

system_terminate(Reason, Parent, Debug, _)

worker_announce(Name, Pid)

workers(Election)

-spec workers(#election{}) -> [node()].

Returns a list of known workers.