pmap (util v1.1.5)

Parallel map and multicall.

Functions

multicall(PidMsgs, Timeout)

-spec multicall([{pid(), term()}], Timeout :: timeout()) ->
             {[OkReply :: term()], [ErrorReply :: term()]}.

Send messages to pids and wait for replies. Each Pid would get a message in the form: {{FromPid, Ref}, Msg} and would have to reply with: FromPid ! {{self(), Ref}, Reply}. The function aggregates all replies into Success and Error lists. The error list is in the form: {Pid, ErrorReason}.

pmap(F, List)

-spec pmap(fun(() -> term()), [Args :: term()]) -> [Reply :: term()].

Evaluate the MultiArgs list by calling F on each argument in the list concurrently.
See also: pmap/3.

pmap(F, Args, Timeout)

-spec pmap(fun((term()) -> term()), Args :: [term()], integer() | infinity) -> [Reply :: term()].

Evaluate the MultiArgs list by calling F on each argument in the list concurrently. Same as pmap/2 but has a Timeout option.

reply(_, Reply)

-spec reply({pid(), reference()}, Reply :: term()) -> ok.

Send a reply back to sender.