prof (util v1.1.5)

Profiling functions

Functions

apply(Fun, Args)

Run a function analysis and save it to "fprof-apply.analysis".

apply(M, F, Args)

Run a function analysis and save it to "fprof-apply.analysis".

start(Pids)

-spec start([pid() | atom()]) -> ok.

Begin profiling a list of pids

stop()

-spec stop() -> ok.

Finish profiling a list of pids and save results to a file called "fprof.analysis".