restrict_remsh_mod (util v1.1.5)

Sample restricted remote shell module disabling q/0 and init:stop/{0,1} commands. The shell introduces a replacement command to stop remote node: remote:stop/1 equivalent to init:stop/1.

To activate restricted shell, run the server node like this: erl -sname node@host +Bi -shell restricted_shell restrict_remsh_mod

Then you can connect to it with: erl -sname a@myhost -remsh node@host
See: http://www.erlang.org/doc/man/shell.html#start_restricted-1

Types

funspec/0

-type funspec() :: {Mod :: atom(), Fun :: atom()}.

Functions

remote_node_stop(Status)

-spec remote_node_stop(Status :: integer()) -> ok.

Replaces init:stop/1 with remote:stop/1 to avoid accidental exit of remote shell.