sntp (util v1.1.5)

Implements SNTP query logic. SNTP - Simple Network Time Protocol (RFC-2030).

Functions

avg_time()

-spec avg_time() -> {Min :: integer(), Max :: integer(), Avg :: integer()}.

Query NTP time sources from "/etc/ntp.conf" and return min/max/avg offset of current host from given time sources.
See also: avg_time/1.

avg_time(ServerAddresses)

-spec avg_time([inet:ip_address()]) -> {Min :: integer(), Max :: integer(), Avg :: integer()}.

Query ServerAddress NTP time sources and return min/max/avg offset of current host from given time sources.

time(ServerAddress)

-spec time(ServerAddress :: inet:ip_address()) -> #sntp{}.

Query ServerAddress time source to find out server time and current host's offset from time source.

time_servers()

-spec time_servers() -> [inet:ip_address()].

Return a list of default NTP time servers for this host.

time_servers(Resolve)

-spec time_servers(boolean()) -> [inet:ip_address()].

Return a list of default NTP time servers for this host. If Resolve is true, the list will contain IP addresses or else host names.