throttle (util v1.1.5)

Throttle given rate over a number of seconds.

Implementation uses time spacing reservation algorithm where each allocation of samples reserves a fraction of space in the throttling window. The reservation gets freed as the time goes by. No more than the Rate number of samples are allowed to fit in the milliseconds Window.
This is an Erlang implementation of the throttling algorithm from the utxx library found at this URL: https://github.com/saleyn/utxx/blob/master/include/utxx/rate_throttler.hpp

Types

throttle/0

-type throttle() :: #throttle{}.

time/0

-type time() :: non_neg_integer().

Functions

add(T)

Add one sample to the throttle

add(T, Samples)

Add Samples to the throttle

add(Throttle, Samples, Now)

-spec add(throttle(), integer(), time()) -> {integer(), throttle()}.

Add Samples to the throtlle. Return {FitSamples, State}, where FitSamples are the number of samples that fit in the throttling window. 0 means that the throttler is fully congested, and more time needs to elapse before the throttles gets reset to accept more samples.

available(T)

See also: available/2.

available(Throttle, Now)

Return the number of available samples given Now current time.

call(T, F)

Call the lambda F, ensuring that it's not called more frequently than the throttle would allow.
Example: 1> T = throttle:new(10, 1000). 2> lists:foldl(fun(_,{T1,A}) -> {T2,R} = throttle:call(T1, fun() -> http:get("google.com") end), {T2, [R|A]} end, {T,[]}, lists:seq(1, 100)).

call(Throttle, F, Now)

-spec call(#throttle{}, fun(() -> any()), non_neg_integer()) -> {#throttle{}, any()}.

Call the lambda F, ensuring that it's not called more throttle would allow.

call(T, M, F, A)

Call M,F,A, ensuring that it's not called more frequently than the throttle would allow.
Example: 1> T = throttle:new(10, 1000). 2> lists:foldl(fun(_,{T1,A}) -> {T2,R} = throttle:call(T1, http, get, ["google.com"]), {T2, [R|A]} end, {T,[]}, lists:seq(1, 100)).

call(Throttle, M, F, A, Now)

-spec call(#throttle{}, atom(), atom(), [any()], non_neg_integer()) -> {#throttle{}, any()}.

Call M,F,A, ensuring that it's not called more frequently than the throttle would allow.

curr_rps(T)

See also: curr_rps/2.

curr_rps(Throttle, Now)

Return currently used rate per second.

new(Rate)

-spec new(non_neg_integer()) -> throttle().

Create a new throttle given the Rate per second.

new(Rate, Window)

-spec new(non_neg_integer(), non_neg_integer()) -> throttle().

See also: new/3.

new(Rate, Window, Now)

-spec new(non_neg_integer(), non_neg_integer(), time()) -> throttle().

Create a new throttle given the Rate per Window milliseconds. Now is expressesed in microseconds since epoch using now().

next_timeout(T)

Return the number of milliseconds to wait until the throttling threshold is satisfied to fit another sample.

now()

reset(Throttle)

See also: reset/2.

reset(Throttle, Now)

Reset the throttle request counter

used(T)

See also: used/2.

used(Throttle, Now)

Return the number of used samples given a_now current time.