user_default (util v1.1.5)

This is an extension of the shell commands to do all the work! Either place this file in the path accessible to Erlang (via ERL_LIBS) or add this line to the ~/.erlang file: code:load_abs(os:getenv("HOME") ++ "/.erlang/user_default").

Functions

dbgadd(Module)

dbgadd(Module, Fun)

dbgdel(Module)

dbgdel(Module, Fun)

dbgoff()

dbgon(Modules)

dbgon(Module, Fun)

dbgtc(File)

debug()

help()

nl()

Load all changed modules on all visible nodes

p(Term)

Term printer

saveh(File)

Save command history to file

tc(N, F)

tc(N, M, F, A)