util_log_color_formatter (util v1.1.5)

Implementation of coloring handler for the Erlang's logger. See https://github.com/hauleth/logger_colorful.

Options

formatter - parent formatter which will be used for message formatting

colors - map of logger:log_level() and colors in form atom describing color name, or 2-ary or 3-ary tuple with 1 or 2 modifiers for the color

Available colors:

black

red

green

yellow

blue

magenta

cyan

white

Available modifiers:

fg - set foreground color

bg - set background color

bright - use bright color

Functions

check_config(Config0)

format(Event, Config0)