Util (Util v1.0.0)

Módulo con funciones que se reutilizan

Autor: Robinson Arias Muñoz.

Fecha: Septiembre del 2026

Licencia: GNU GPL v3

Functions

ingresar(mensaje, atom)

Función para ingresar un dato (:texto, :entero o :real) desde el teclado

Parámetros

mensaje, texto que se le presenta al usuario

Ejemplos

  iex> Util.ingresar("Ingrese el nombre: ", :texto)

  iex> Util.ingresar("Ingrese la edad: ", :entero)

  iex> Util.ingresar("Ingrese la altura: ", :real)

o puede usar

  "Ingrese el nombre: "
  |> Util.ingresar(:texto)

  "Ingrese la edad: "
  |> Util.ingresar(:entero)

  "Ingrese la altura: "
  |> Util.ingresar(:real)

ingresar_texto(mensaje)

Por comparabilidad con el ejemplo v2, pero será descontinuada desde al v3.

mostrar_error(mensaje)

Función para mostrar un mensaje de error en la pantalla.

Parámetros

mensaje, texto que se le presenta al usuario como error

Ejemplos

  iex> Util.mostrar_error("Dato inválido")

o puede usar

  "Dato inválido"
  |> Util.mostrar_error()

mostrar_mensaje(mensaje)

Función para mostrar un mensaje en la pantalla.

Parámetros

mensaje, texto que se le presenta al usuario

Ejemplos

  iex> Util.mostrar_mensaje("Hola Mundo")

o puede usar

  "Hola Mundo"
  |> Util.mostrar_mensaje()